Improve helm class reusing code
