+++ tests/test_jinja2.py
@@ -13,10 +13,11 @@ import time
 import unittest
 
 from kapitan import cached
-from kapitan.jinja2_filters import base64_encode, render_jinja2_file
+from kapitan.jinja2_filters import base64_encode
 from kapitan.refs.base import RefController, Revealer
 from kapitan.refs.base64 import Base64Ref
 from kapitan.resources import inventory
+from kapitan.utils import render_jinja2_file
 
 
 class Jinja2FiltersTest(unittest.TestCase):
